split out dashboard files
